Allow control over method name used in side navigation.
If spec.x-methods-by-type is true, then method type (GET, POST...) will be used in navigation.
If operation.x-navigation-name is set, then overrides the navigation name for the method.
This allows pure REST APIs to have the methods for an API endpoint grouped by a TAG, and each
listed in the navigation by type (get, post etc). In the case where a GET method returns a list
x-navigation-name can be set to be "list' to differentiate from a single resource GET.
Other usecases and permutations are possible.